streamlit_wordcloud/requirements.txt
streamlit
matplotlib
wordcloud
numpy
Pillow
requests
beautifulsoup4
